Extract client component from HomePage
